services/resume-parser/pyproject.toml
[project]
name = "resume-parser"
version = "0.1.0"
description = "AI-powered resume parsing microservice"
authors = [
    {name = "Sashvad (Sachin) Satish Kumar",email = "sskumar@umd.edu"}
]
license = {text = "N"}
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (==0.104.1)",
    "uvicorn (==0.24.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-multipart (==0.0.6)",
    "pypdf2 (==3.0.1)",
    "python-docx (==1.1.0)",
    "spacy (==3.7.2)",
    "structlog (==23.2.0)",
    "httpx (==0.25.2)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"
pytest-asyncio = "0.21.1"
pytest-cov = "4.1.0"
black = "23.11.0"
isort = "5.12.0"
flake8 = "6.1.0"
mypy = "1.7.1"

